requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
Django==2.1.8
django-allauth==0.39.1
django-appconf==1.0.3
django-bootstrap-pagination==1.7.0
django-crispy-forms==1.7.2
django-imagekit==4.0.2
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
oauthlib==3.0.1
pilkit==2.0
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.3
Werkzeug==0.15.2
